requirements/reQ.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-analytical==2.4.0
django-anymail==3.0
django-appconf==1.0.2
django-ckeditor==5.6.1
django-compressor==2.2
django-crispy-forms==1.7.2
django-environ==0.4.5
django-js-asset==1.1.0
django-redis==4.9.0
django-taggit==0.22.2
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
more-itertools==4.3.0
oauthlib==2.1.0
Pillow==5.2.0
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.5.4
pytest==3.7.0
python-slugify==1.2.5
python3-openid==3.1.0
pytz==2018.5
rcssmin==1.0.6
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rjsmin==1.0.12
six==1.11.0
Unidecode==1.0.22
urllib3==1.23
uWSGI==2.0.17.1
Werkzeug==0.14.1
